GraphsOptim

Documentation for GraphsOptim.

Index

GraphsOptim.GraphsOptim

GraphsOptim.minimum_cost_flow

GraphsOptim.minimum_cost_flow!

API Reference

GraphsOptim.GraphsOptim — Module

GraphsOptim

A package for graph-related optimization algorithms that rely on Linear Programming.

GraphsOptim.minimum_cost_flow — Function

minimum_cost_flow(
    g, vertex_demand, edge_cost, edge_min_capacity, edge_max_capacity;
    flowvar_name, integer
)

Create a JuMP model, call minimum_cost_flow! to populate it, optimize it and return the optimal flow as a sparse matrix.

GraphsOptim.minimum_cost_flow! — Method

minimum_cost_flow!(
    model, g, vertex_demand, edge_cost, edge_min_capacity, edge_max_capacity;
    flowvar_name, integer
)

Modify a JuMP model by adding the variables, constraints and objective necessary to compute a minimum cost flow over a directed graph.

Arguments

model::JuMP.Model: the optimization model to modify

g::Graphs.AbstractGraph: a directed graph $G = (V, E)$

vertex_demand::AbstractVector: a vector $d \in \mathbb{R}^V$ giving the flow requested by each vertex (should be positive for sinks, negative for sources and zero elsewhere)

edge_cost::AbstractMatrix: a vector $c \in \mathbb{R}^E$ giving the cost of a unit of flow on each edge

edge_min_capacity::AbstractMatrix: a vector $a \in \mathbb{R}^E$ giving the minimum flow allowed on each edge

edge_max_capacity::AbstractMatrix: a vector $b \in \mathbb{R}^E$ giving the maximum flow allowed on each edge

Keyword arguments

flowvar_name::Symbol: the name of the optimization variable containing the flow

integer::Bool: whether the flow should be integer-valued or real-valued

Mathematical formulation

The objective function is

\[\min_{f \in \mathbb{R}^E} \sum_{(u, v) \in E} c(u, v) f(u, v)\]

The edge capacity constraint dictates that for all $(u, v) \in E$,

\[a(u, v) \leq f(u, v) \leq b(u, v)\]

The flow conservation constraint with node demand dictates that for all $v \in V$,

\[f^-(v) = d(v) + f^+(v)\]

where the incoming flow $f^-(v)$ and outgoing flow $f^+(v)$ are defined as

\[f^-(v) = \sum_{u \in N^-(v)} f(u, v) \quad \text{and} \quad f^+(v) = \sum_{w \in N^+(v)} f(v, w)\]