GraphsOptim

Documentation for GraphsOptim.

Index

GraphsOptim.GraphsOptim

GraphsOptim.MatchingResult

GraphsOptim.cutoff_weights

GraphsOptim.default_weights

GraphsOptim.maximum_weight_matching

GraphsOptim.maximum_weight_maximal_matching

GraphsOptim.minimum_cost_flow

API Reference

GraphsOptim.GraphsOptim — Module

GraphsOptim

A package for graph-related optimization algorithms that rely on Linear Programming.

The package JuMP.jl and one of its supported solvers are required.

GraphsOptim.MatchingResult — Type

MatchingResult{U}

A type representing the result of a matching algorithm.

Fields

weight::U: total weight of the matching

mate::Vector{Int}: pairwise assignment.

mate[i] = j if vertex i is matched to vertex j, and mate[i] = -1 for unmatched vertices.

GraphsOptim.cutoff_weights — Method

cutoff_weights(w, cutoff)

Copy the weights matrix w with all elements below cutoff set to 0.

GraphsOptim.default_weights — Method

default_weights(g)

Return the binary undirected adjacency matrix of g.

GraphsOptim.maximum_weight_matching — Method

maximum_weight_matching(g, w; optimizer)

Given a graph g and an matrix w of edge weights, return a matching (MatchingResult object) with the maximum total weight.

If no weight matrix is given, all edges will be considered to have weight 1 (results in max cardinality matching). Edges in g that are not present in w will not be considered for the matching.

A JuMP-compatible solver must be provided with the optimizer argument.

The efficiency of the algorithm depends on the input graph:

If the graph is bipartite, then the LP relaxation is integral.

If the graph is not bipartite, then it requires a MIP solver and the computation time may grow exponentially.

GraphsOptim.maximum_weight_maximal_matching — Method

maximum_weight_maximal_matching(g, w[, cutoff]; optimizer)

Given a bipartite graph g and a matrix w of edge weights, return a matching (MatchingResult object) with the maximum total weight among the ones that contain the largest number of edges.

If no weight matrix is given, all edges will be considered to have weight 1 (results in max cardinality matching). Edges in g that are not present in w will not be considered for the matching.

A JuMP-compatible solver must be provided with the optimizer argument.

A cutoff argument can be given to reduce computation time by excluding edges with weights lower than the specified value.

The algorithm relies on a linear relaxation on of the matching problem, which is guaranteed to have integer solution on bipartite graphs.

GraphsOptim.minimum_cost_flow — Method

minimum_cost_flow(
    g::AbstractGraph,
    node_demand::AbstractVector,
    edge_capacity::AbstractMatrix,
    edge_cost::AbstractMatrix,
    optimizer;
    edge_demand::Union{Nothing,AbstractMatrix} = nothing,
    source_nodes = (),
    sink_nodes = ()
)

Find a flow over a directed graph g satisfying the node_demand at each node and edge_capacity constraints for each edge while minimizing dot(edge_cost, flow).

Returns a sparse flow matrix, flow[i,j] corresponds to the flow on the (i,j) arc.

Arguments

g is a directed Graphs.AbstractGraph.

-node_demand is a vector of nodal demand values, which should be negative for sources, positive for sink nodes, and zero for all other nodes.

edge_capacity::AbstractMatrix sets an upper bound on the flow of each arc.

edge_cost::AbstractMatrix the cost per unit of flow on each arc.

optimizer is an optimizer constructor, like Clp.Optimizer or () -> Clp.Optimizer() passed at the construction of the JuMP.Model.

Keyword arguments

edge_demand::Union{Nothing,AbstractMatrix}: require a minimum flow for edges, or nothing.

source_nodes Collection of sources at which the nodal netflow is allowed to be greater than nodal demand, defaults to an empty tuple.

sink_nodes Collection of sinks at which the nodal netflow is allowed to be less than nodal demand, defaults to an empty tuple.

source_nodes & sink_nodes are only needed when nodal flow are not explictly set in node_demand

Examples

julia> import Graphs
julia> using GraphsFlows: mincost_flow
julia> import GLPK # use your favorite LP solver here
julia> using SparseArrays: spzeros
julia> g = Graphs.DiGraph(6) # Create a flow-graph
julia> Graphs.add_edge!(g, 5, 1)
julia> Graphs.add_edge!(g, 5, 2)
julia> Graphs.add_edge!(g, 3, 6)
julia> Graphs.add_edge!(g, 4, 6)
julia> Graphs.add_edge!(g, 1, 3)
julia> Graphs.add_edge!(g, 1, 4)
julia> Graphs.add_edge!(g, 2, 3)
julia> Graphs.add_edge!(g, 2, 4)
julia> cost = zeros(6,6)
julia> cost[1,3] = 10
julia> cost[1,4] = 5
julia> cost[2,3] = 2
julia> cost[2,4] = 2
julia> demand = spzeros(6)
julia> demand[5] = -2
julia> demand[6] = 2
julia> capacity = ones(6,6)
julia> flow = minimum_cost_flow(g, demand, capacity, cost, optimizer=GLPK.Optimizer)